# pkg/latex/docs/swagger.yaml
basePath: /api/doc
info:
  contact:
    email: support@bullteam.cn
    name: API Support
  description: TEFS latex
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TEFS latex API
  version: V0.1
paths:
  /latex:
    get:
      consumes:
      - application/json
      description: 获取latex列表
      parameters:
      - description: 状态 1=正常 2=归档 3=删除
        in: query
        name: status
        required: true
        type: integer
      - description: 从第多少条起始
        in: query
        name: offset
        required: true
        type: integer
      - description: 获取条数
        in: query
        name: limit
        required: true
        type: integer
      - description: 创建时间 create_at; 最后更新时间update_at; 当前用户操作时间current_user_operation_time;
        in: query
        name: sort_field
        required: true
        type: string
      - description: asc升序 desc降序
        in: query
        name: sort
        required: true
        type: string
      - description: latex文档ID(非必填项)
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取latex列表
      tags:
      - latex
    post:
      consumes:
      - application/json
      description: 创建latex
      parameters:
      - description: latex 名字
        in: query
        name: name
        type: string
      - description: 创建latex类型 1=空白创建(只会生成main.txt文件);2=样例创建;3=上传创建(zip文件);4=模板创建
        in: query
        name: create_latex_type
        required: true
        type: string
      - description: create_latex_type=3(上传创建)时的zip文件(只支持zip文件)。
        in: formData
        name: zip_file
        type: file
      - description: create_latex_type=4(模板创建)时的模板名称
        in: query
        name: template_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建latex
      tags:
      - latex
    put:
      consumes:
      - application/json
      description: latex批量修改,单个修改(归档,丢弃到垃圾桶,删除)
      parameters:
      - description: latex ID列表(只要option=3的时 并且该latex为用户创建时,需要此参数)，以,分隔 例如 1,2,3
        in: query
        name: latex_ids
        type: string
      - description: user latex ID列表，以,分隔 例如 1,2,3 (只要option=3的时, 并且该latex不是用户创建，放入此参数。表示用户要离开该latex)
        in: query
        name: user_latex_ids
        type: string
      - description: 1=恢复到正常状态;2=归档;3=丢弃到回收站;4=删除
        in: query
        name: option
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: latex批量修改,单个修改(归档,丢弃到垃圾桶,删除)
      tags:
      - latex
  /latex/{id}:
    delete:
      consumes:
      - application/json
      description: 删除latex
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除latex
      tags:
      - latex
    put:
      consumes:
      - application/json
      description: 修改latex
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: latex 名字
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改latex
      tags:
      - latex
  /latex/{id}/config:
    get:
      consumes:
      - application/json
      description: 获取latex配置信息
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取latex配置信息
      tags:
      - latex
    put:
      consumes:
      - application/json
      description: 修改latex配置信息
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: 编译文件路径
        in: query
        name: main_document
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改latex配置信息
      tags:
      - latex
...
  /latex_templates:
    get:
      consumes:
      - application/json
      description: 获取latex模板
      parameters:
      - description: 从多少条开始
        in: query
        name: offset
        required: true
        type: integer
      - description: 返回的条数
        in: query
        name: limit
        required: true
        type: integer
      - description: 获取模板类型：paper=论文模板，notes=笔记模板
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取latex模板
      tags:
      - template
    post:
      consumes:
      - application/json
      description: 上传latex模板(需要管理员权限)
      parameters:
      - description: 模板作者
        in: query
        name: author
        required: true
        type: string
      - description: 模板license
        in: query
        name: license
        required: true
        type: string
      - description: 模板摘要
        in: query
        name: abstract
        required: true
        type: string
      - description: 模板封面图
        in: formData
        name: image
        required: true
        type: file
      - description: 要操作的文件
        in: formData
        name: file
        required: true
        type: file
      - description: 上传模板类型：paper=论文模板，notes=笔记模板
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 上传latex模板(需要管理员权限)
      tags:
      - template
  /latex_templates/{id}:
    delete:
      consumes:
      - application/json
      description: 删除latex模板(需要管理员权限)
      parameters:
      - description: 要操作的文件
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除latex模板(需要管理员权限)
      tags:
      - template
    put:
      consumes:
      - application/json
      description: 更新latex模板(需要管理员权限)
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: 模板作者
        in: query
        name: author
        type: string
      - description: 模板license
        in: query
        name: license
        type: string
      - description: 模板摘要
        in: query
        name: abstract
        type: string
      - description: 模板封面图
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新latex模板(需要管理员权限)
      tags:
      - template
...
  /tag/{id}/latex/{latex_id}:
    delete:
      consumes:
      - application/json
      description: tag添加latex
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: latex ID
        in: path
        name: latex_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: latex删除tag
      tags:
      - latexTag
    post:
      consumes:
      - application/json
      description: tag添加latex
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: latex ID
        in: path
        name: latex_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: latex添加tag
      tags:
      - latexTag
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"